requirements.txt
beautifulsoup4>=4.12.2
markdownify>=0.11.6
requests>=2.30.0
PyYAML==6.0
